mpd-DATPERF-2692-implicit-etl-refactor

See merge request thetradedesk/teams/aifun/mlplatform!416

+++ plutus/src/main/scala/com/thetradedesk/plutus/data/schema/DiscrepancyDataset.scala
@@ -6,7 +6,7 @@ final case class Pda(SupplyVendorName: String, PartnerId: String, DiscrepancyAdj
 
 final case class Deals(SupplyVendorId: String, SupplyVendorDealCode: String, IsVariablePrice: Boolean)
 
-final case class EmpiricalDiscrepancy(PartnerId: String, SupplyVendor: String, DealId: String, AdFormat: String, EmpiricalDiscrepancy: BigDecimal)
+final case class EmpiricalDiscrepancy(PartnerId: String, SupplyVendor: String, DealId: String, AdFormat: String, EmpiricalDiscrepancy: Double)
 
 object DiscrepancyDataset {
